% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPalette.R
\name{genPalette}
\alias{genPalette}
\title{Generate a colour palette for the wmts}
\usage{
genPalette(brk, crp, lowerTransparent = TRUE)
}
\arguments{
\item{brk}{list of break values between classes, including the upper and lower limits}

\item{crp}{a colorRampPalette used for generating the colors for each class}

\item{lowerTransparent}{logical - should the lowest class be transparent}
}
\description{
Generate a colour palette for the wmts
}
\details{
Generates a data frame of breaks and Hex colour codes. Values above the upper limit are treated as having colour NA
}
