% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTiles.R
\name{genTiles}
\alias{genTiles}
\title{Function to generate tiles for 0.1 degree data}
\usage{
genTiles(
  rst,
  fcol,
  outdir = ".",
  minZ = 0,
  maxZ = 3,
  w = 512,
  h = w,
  lv = NULL,
  ...
)
}
\arguments{
\item{rst}{a SpatRaster or file name to open as a Spatraster to generate tiles from}

\item{fcol}{function to convert raster values into an RGBA matrix - see details}

\item{outdir}{directory into which the tiles are written}

\item{minZ}{lowest zoom level to generate}

\item{maxZ}{maximum zoom level to generate}

\item{w}{width of tiles in pixels}

\item{h}{height of tiles in pixels}

\item{lv}{values used to generage a legend. If \code{NULL} to legend is generated.}

\item{...}{passed to terra project or terra resample}
}
\description{
Function to generate tiles for 0.1 degree data
}
\details{
This is naive generater.
Defaults work well for plotting 0.1 degree data in a limited number of tiles.
fcol should take a vector of n values and return an n hex colour values which should include transparency.
}
