% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPalette.R
\name{plotPalette}
\alias{plotPalette}
\title{Function for Generating a legend from a palette}
\usage{
plotPalette(pal, fileName = "legend.svg")
}
\arguments{
\item{pal}{the Palette to plot}

\item{fileName}{location of the svg image legend}
}
\description{
Function for Generating a legend from a palette
}
