% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveTiles.R
\name{serveTiles}
\alias{serveTiles}
\title{Start a serving serving tiles from a raster data set with time attribute}
\usage{
serveTiles(
  dataFile,
  fcol,
  lv = NULL,
  cachePath = NULL,
  staticPath = NULL,
  fplot = function(t, v) {     plot(t, v) },
  ferror = function(err) {     err$message },
  port = NULL,
  clearCache = FALSE
)
}
\arguments{
\item{dataFile}{file containing spatial data to be read by \code{terra::rast}}

\item{fcol}{function for returning colours}

\item{lv}{values for drawing in legend}

\item{cachePath}{path to directory for storing cached files}

\item{staticPath}{path for serving static assets from}

\item{fplot}{function to plot a timeseries of data (see details)}

\item{port}{port to serve data on}

\item{clearCache}{logical, should cache be cleared if file changes}

\item{ferr}{function to process errors and return messages (see details)}
}
\description{
Start a serving serving tiles from a raster data set with time attribute
}
\details{
For a decription of fcol and lv see \code{scaleImage}. fplot shoudl take as its first input a POSIXct vector and as for its second a numeric vector of the same length and return a plot. ferror creates the response body contents from the standard R err object.
}
