% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleTile.R
\name{singleTile}
\alias{singleTile}
\title{Function to generate a single tile}
\usage{
singleTile(rst, pal, Z, X, Y, w = 512, h = w, method = c("bilinear", "ngb"))
}
\arguments{
\item{rst}{a SpatRaster object to generate tiles from}

\item{pal}{a palette of breaks and colours as generated by \code{genPalette}}

\item{Z}{zoom level}

\item{X}{value of the tile}

\item{Y}{value of the tile}

\item{w}{width of tiles in pixels}

\item{h}{height of tiles in pixels}

\item{method}{method of interpolation (not used)}
}
\value{
a raw png vector
}
\description{
Function to generate a single tile
}
\details{
This uses the terra library to project \code{rst} onto the correct tile area. Tiles are then generated by cutting into the classes in pal.
}
