% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveWMTS.R
\name{wmtsServer}
\alias{wmtsServer}
\title{R6 Class for wmts}
\description{
R6 Class for wmts

R6 Class for wmts
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wmtsServer-new}{\code{wmtsServer$new()}}
\item \href{#method-wmtsServer-static}{\code{wmtsServer$static()}}
\item \href{#method-wmtsServer-start}{\code{wmtsServer$start()}}
\item \href{#method-wmtsServer-overload}{\code{wmtsServer$overload()}}
\item \href{#method-wmtsServer-clone}{\code{wmtsServer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wmtsServer-new"></a>}}
\if{latex}{\out{\hypertarget{method-wmtsServer-new}{}}}
\subsection{Method \code{new()}}{
Creates a wmtsServer object from a wmts object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wmtsServer$new(wmts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wmts}}{a wmts object}

\item{\code{port=NULL}}{port to serve the data on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(self) suitable for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wmtsServer-static"></a>}}
\if{latex}{\out{\hypertarget{method-wmtsServer-static}{}}}
\subsection{Method \code{static()}}{
Shows, or sets the staticPath value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wmtsServer$static(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{see details}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If \code{x=NULL} the current value is returned. If \code{x=NA} then no path is set and caching does not occur (the default setup).
If x can be interpreded as a character vector then the path is set (subject to it existing)
}

\subsection{Returns}{
the value taken by cachePath at the end of the call
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wmtsServer-start"></a>}}
\if{latex}{\out{\hypertarget{method-wmtsServer-start}{}}}
\subsection{Method \code{start()}}{
Start the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wmtsServer$start(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to plumber::pr_run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wmtsServer-overload"></a>}}
\if{latex}{\out{\hypertarget{method-wmtsServer-overload}{}}}
\subsection{Method \code{overload()}}{
Shows or Sets one over the overloadable function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wmtsServer$overload(f = NULL, name = c("fError", "fCORS"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{see details}

\item{\code{name}}{name of function to overload}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If \code{f=NULL} the current value of the names function is returned.
If \code{f=NA} then the default function is set. Otherwise if f is a function it set to be used.
The function being overloaded should take the following inputs and outputs

fError
Input: an R error
Output: Message to display
}

\subsection{Returns}{
Either the current value or invisible(self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wmtsServer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wmtsServer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wmtsServer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
