% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTiles.R
\name{genTiles}
\alias{genTiles}
\title{Function to generate tiles for 0.1 degree data}
\usage{
genTiles(
  rst,
  pal,
  outdir = ".",
  maxZ = 3,
  w = 445,
  h = 223,
  mapCRS = "EPSG:4326"
)
}
\arguments{
\item{rst}{a raster to generate tiles from}

\item{pal}{a palette of breaks and colours as generated by \code{genPalette}}

\item{outdir}{directory into which the tiles are written}

\item{maxZ}{mximum zoom level to generate}

\item{w}{width of tiles in pixels}

\item{h}{height of tiles in pixels}

\item{mapCRS}{CRS (as a string) for the generated tiles}
}
\description{
Function to generate tiles for 0.1 degree data
}
\details{
This is naive implimtnation. It scales or reprojects raster so dimensions are
(2^maxZ)*w and (2^maxZ)*h.
Tiles are then generated by aggregating to (2^Z)*w and (2^Z)*h then cutting into the classes in pal.
Defaults work well for plotting 0.1 degree data in a limited number of tiles
}
