% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTiles.R
\name{genTiles}
\alias{genTiles}
\title{Function to generate tiles for 0.1 degree data}
\usage{
genTiles(
  rst,
  pal,
  outdir = ".",
  maxZ = 3,
  w = 512,
  h = w,
  method = c("bilinear", "ngb")
)
}
\arguments{
\item{rst}{a SpatRaster or file name to open as a Spatraster to generate tiles from}

\item{pal}{a palette of breaks and colours as generated by \code{genPalette}}

\item{outdir}{directory into which the tiles are written}

\item{maxZ}{mximum zoom level to generate}

\item{w}{width of tiles in pixels}

\item{h}{height of tiles in pixels}

\item{method}{- not used}
}
\description{
Function to generate tiles for 0.1 degree data
}
\details{
This is naive generater.
Defaults work well for plotting 0.1 degree data in a limited number of tiles
}
