% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleImage.R
\name{scaleImage}
\alias{scaleImage}
\title{Function for Generating a legend from fcol}
\usage{
scaleImage(fcol, lv, nb = 0, w = 5, h = 10)
}
\arguments{
\item{h}{}
}
\description{
Function for Generating a legend from fcol
}
