% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleTile.R
\name{singleTile}
\alias{singleTile}
\title{Function to generate a single tile}
\usage{
singleTile(rst, fcol, Z, X, Y, w = 512, h = w, ...)
}
\arguments{
\item{rst}{a SpatRaster object to generate tiles from}

\item{fcol}{function to convert raster values into an hex colour scheme}

\item{Z}{zoom level}

\item{X}{value of the tile}

\item{Y}{value of the tile}

\item{w}{width of tiles in pixels}

\item{h}{height of tiles in pixels}

\item{...}{passed to terra project or terra resample}
}
\value{
a raw vector containing the png image
}
\description{
Function to generate a single tile
}
\details{
This uses the terra library to project \code{rst} onto the correct tile area. Tiles are then generated by applying the function fcol. fcol should take a vector of n values and return an n hex colour values which should include transparency.
}
