% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmts.R
\name{wmts}
\alias{wmts}
\title{OGC Webmap Tile Server (WMTS)}
\usage{
wmts(x, loc, buffer = NULL, ..., zoom = NULL, max_tiles = 25,
  bands = 1:3)
}
\arguments{
\item{x}{WMTS address}

\item{loc}{something we can get an extent from}

\item{buffer}{radius in metres (in the case 'loc' is a single point)}

\item{...}{ignored}

\item{zoom}{override hueristic for zoom level (take care! no safety checks)}

\item{max_tiles}{max number of tiles for zoom hueristic (default 25)}

\item{bands}{bands to get (default is 1,2,3 assuming RGB)}
}
\value{
raster brick
}
\description{
Get raster from WMTS
}
\examples{
centre <- c(-9004415, 3806965)
radius <- 4000
u <- "WMTS:https://basemap.nationalmap.gov/arcgis/rest/services/USGSTopo/MapServer/WMTS/1.0.0/WMTSCapabilities.xml,layer=USGSTopo,tilematrixset=default028mm"
x <- wmts(u, centre, radius)
#ex <- raster::extent(centre[c(1, 1, 2, 2)] + c(-1, 1, -1, 1) * buffer)
#rgb_map <- plot_get_wmts(l, ex, native = TRUE)
#
#library(ggplot2)
#
#tab <- tibble::as_tibble(as.data.frame(rgb_map, xy = TRUE))
#names(tab) <- c("x", "y", "red", "green", "blue")
##tab <- dplyr::filter(tab, !is.na(red))
#
### ... when we have missing values, we should drop them or rgb() will error
#tab$hex <- rgb(tab$red, tab$green, tab$blue, maxColorValue = 255)
#ggplot(tab, aes(x, y, fill = hex)) +
#  geom_raster() +
#  coord_equal() +
#  scale_fill_identity()
}
